refactored into package newick.node
