+++ src/main/java/nl/tudelft/context/drawable/NewickLabel.java
@@ -1,8 +1,8 @@
 package nl.tudelft.context.drawable;
 
 import javafx.scene.control.Label;
-import nl.tudelft.context.model.newick.AbstractNode;
-import nl.tudelft.context.model.newick.AncestorNode;
+import nl.tudelft.context.model.newick.node.AbstractNode;
+import nl.tudelft.context.model.newick.node.AncestorNode;
 import nl.tudelft.context.model.newick.selection.Selection;
 
 /**
