+++ src/test/java/nl/tudelft/context/controller/NewickControllerTest.java
@@ -5,8 +5,8 @@ import de.saxsys.javafx.test.JfxRunner;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.scene.control.MenuBar;
 import javafx.scene.control.MenuItem;
-import nl.tudelft.context.model.newick.AbstractNode;
-import nl.tudelft.context.model.newick.LeaveNode;
+import nl.tudelft.context.model.newick.node.AbstractNode;
+import nl.tudelft.context.model.newick.node.LeaveNode;
 import nl.tudelft.context.model.newick.Newick;
 import nl.tudelft.context.model.newick.selection.All;
 import nl.tudelft.context.workspace.Workspace;
